+++ R/c4a.R
@@ -2,11 +2,11 @@
 #'
 #' Get a cols4all color palette: `c4a` returns the colors of the specified palette, `c4a_na` returns the color for missing value that is associated with the specified palette, and `c4a_ramp` returns a color ramp function.  Run \code{\link{c4a_gui}} to see all available palettes, which are also listed with \code{\link{c4a_palettes}}.
 #'
-#' @param palette name of the palette. See \code{\link{c4a_palettes}} for available palettes. If omitted, the default palette is provided by `c4a_default_palette`. The palette name can be prefixed with a `"-"` symbol, which will reverse the palette (this can also be done with the `reverse` argument).
+#' @param palette name of the palette. See \code{\link{c4a_palettes}} for available palettes. If omitted, the default palette is provided by `c4a_default_palette`. The palette name can be prefixed with a `"-"` symbol, which will reverse the palette (this can also be done with the `reverse` argument). For bivariate palettes, a `"-"` means reversed horizontally (columns), a `"|"`means reversed vertically (row), and a `"+"` means reversed in both directions.
 #' @param n number of colors. If omitted then: for type `"cat"` the maximum number of colors is returned, for types `"seq"`, `"div"`, and `"cyc"`, 7 , 9, and 9 colors respectively.
 #' @param m number of rows in case type is bivariate, so one of `"bivs"`, `"bivc"`, `"bivd"` or  `"bivg"` (see \code{\link{c4a_types}} for descriptions)
 #' @param type type of color palette, in case `palette` is not specified: one of `"cat"`, `"seq"`, `"div"`, `"cyc"`, `"bivs"`, `"bivc"`, `"bivd"`, `"bivg"`. Run \code{\link{c4a_types}} for descriptions.
-#' @param reverse should the palette be reversed?
+#' @param reverse should the palette be reversed? In case of a bivariate palette, a vector of two: the first indicates the horizontal direction (columns) and the second the vertical (rows).
 #' @param order order of colors. Only applicable for `"cat"` palettes
 #' @param range a vector of two numbers between 0 and 1 that determine the range that is used for sequential and diverging palettes. The first number determines where the palette begins, and the second number where it ends. For sequential `"seq"` palettes, 0 means the leftmost (normally lightest) color, and 1 the rightmost (often darkest) color. For diverging `"seq"` palettes, 0 means the middle color, and 1 both extremes. If only one number is provided, this number is interpreted as the endpoint (with 0 taken as the start).
 #' @param colorsort Sort the colors. Options: `"orig"` (original order), `"Hx"` (hue, where x is a starting number from 0 to 360), `"C"` (chroma), `"L"` (luminance). All these options are available for `"cat"` palettes, only the last one for `"seq"`, and none for the other palette types.
@@ -53,6 +53,8 @@ c4a = function(palette = NULL, n = NA, m = NA, type = c("cat", "seq", "div", "cy
 	type = x$type
 
 	if (is.null(x)) return(invisible(NULL))
+
+	reverse = rep(reverse, length.out = 2)
 
 	reverse = xor(reverse, x$reverse)
 
@@ -102,7 +104,21 @@ c4a = function(palette = NULL, n = NA, m = NA, type = c("cat", "seq", "div", "cy
 	} else pal
 
 	if (!is.null(mes) && verbose) message(mes)
-	pal2 = if (reverse) rev(pal) else pal
+
+	if (substr(type, 1, 3) == "biv") {
+		if (reverse[1]) {
+			pal2 = pal[, ncol(pal):1L]
+		} else {
+			pal2 = pal
+		}
+
+		if (reverse[2]) {
+			pal2 = pal2[nrow(pal2):1L, ]
+		}
+	} else {
+		pal2 = if (reverse[1]) rev(pal) else pal
+	}
+
 
 	if (format == "hex") {
 		pal2
@@ -150,8 +166,11 @@ c4a_info = function(palette, no.match = c("message", "error", "null"), verbose =
 	if (length(palette) != 1L) stop("palette should be a character value (so length 1)", call. = FALSE)
 
 	no.match = match.arg(no.match)
-	isrev = (substr(palette, 1, 1) == "-")
-	if (isrev) palette = substr(palette, 2, nchar(palette))
+	isrev = (substr(palette, 1, 1) %in% c("-", "+"))
+	isrev2 = (substr(palette, 1, 1) %in% c("|", "+"))
+	if (isrev || isrev2) {
+		palette = substr(palette, 2, nchar(palette))
+	}
 
 	z = .C4A$z
 
@@ -162,7 +181,7 @@ c4a_info = function(palette, no.match = c("message", "error", "null"), verbose =
 	palid = which(fullname == z$fullname)
 
 	x = as.list(z[palid, ])
-	x$reverse = isrev
+	x$reverse = c(isrev, isrev2)
 	x$palette = x$palette[[1]]
 
 	structure(x, class = c("c4a_info", "list"))