+++ NEWS.md
@@ -1,3 +1,6 @@
+# cols4all 0.8-1 (devel)
+- added "|" and "+" prefixes to reverse bivariate palettes
+
 # cols4all 0.8
 - main table much faster
 - sorting and filtering options added